add rules of thumb file
